apps/provider/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.0
aiosignal==1.3.1
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
black==24.1.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
commit-linter==1.0.3
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.1.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
environ==1.0
frozenlist==1.5.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
Markdown==3.5.2
msgpack==1.0.8
multidict==6.1.0
mypy==1.9.0
mypy-extensions==1.0.0
mypy1989==0.0.2
numpy==2.1.3
packaging==23.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2023.3.post1
pyxlsb==1.0.10
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
twilio==9.3.6
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
zope.interface==6.2
